Use XLS template for plant lines instead of CSV
