+++ config/locales/errors.en.yml
@@ -16,7 +16,9 @@ en:
           attributes:
             file:
               no_header: 'has no correct header provided. At least four columns are expected.'
-              no_header_plant_lines: 'has no correct header provided.'
+              no_plant_lines_sheet: "contents invalid. 'Plant lines' sheet missing."
+              plant_lines_parsing_error: "could not be parsed. An unknown error occured: %{error}. Please make sure to use an actual XLS file."
+              invalid_plant_lines_header: "contents invalid. Header rows seem to be altered."
               no_plant_accession_header: 'has no correct header provided. Please provide the "Plant accession" column.'
               no_originating_organisation_header: 'has no correct header provided. Please provide the "Originating organisation" column.'
               no_year_produced_header: 'has no correct header provided. Please provide the "Year produced" column.'
