--- public/manuals/plant_lines_template.csv
@@ -1,2 +1,0 @@
-"Species","Plant variety","Crop type","Plant line","Common name","Previous line name","Genetic status","Sequence","Plant accession","Originating organisation","Year produced"
-"Species – replace it","Plant variety – replace it","Crop type - replace it","Plant line – replace it","Common name - replace it","Previous line name - replace it","Genetic status - replace it","Sequence - replace it","Plant accession - replace it","Originating organisation - replace it","Year produced - replace it"
